A fair bit of work has gone into making the editing UI look reasonably consistent for editing projects, sites, datasets and records. This includes:
Using the InputTextModule to make text fields look consistent with other fields.

Same grid layout as much as possible.

All edit pages are now responsive and inputs are wider.

All inputs are consistent with the theme.

Buttons across bottom - delete, cancel and save - are consistent as much as possible.

There is error feedback for all fields.

Also added comments field for projects which show in project lists and render as html.

Fixed problem with not being able to click records table to edit.